# promql: more Kahan summation (avg) and less incremental mean calculation (avg, avg_over_time)