See various commits for details.

Studying #14074, I noticed that the tests added for the `sum` aggregation wouldn't work for the `avg` aggregation. Also, the tests already there for `sum_over_time` (with a series `1 1e100 1 -1e100`) wouldn't work for `avg_over_time`.

Simply applying Kahan summation wouldn't help. (In fact, it was already applied to `avg_over_time`, but of course, without adding the test mentioned above.) The reason is that we always use incremental calculation of the mean value. That's a good approach if the sum of the averaged values overflow float64. But in general, it's just more costly and has way more potential of accumulating numerical errors.

This PR changes the behavior that incremental calculation of the mean is only used if the sum would overflow. It also uses Kahan summation in all cases, which should take care of errors happening when adding small numbers to large ones (which is a case where incremental calculation of the mean might help, too, but I believe Kahan summation is cheaper and better and also consistent with how we calculate sums).