@beorn7 2024-07-04
BTW, the speedup for `avg` calculation isn't significant, but it should be noted that this PR adds Kahan summation to `avg`, which has lead to a measurable slowdown for `sum` in #14074. Or in other words, the slowdown caused by the Kahan summation is approximately compensated by the speedup gained from removing incremental mean calculation.

```
name                                          old time/op    new time/op    delta
RangeQuery/expr=avg(a_hundred),steps=1-16        224µs ± 2%     216µs ± 1%  -3.39%  (p=0.000 n=10+9)
RangeQuery/expr=avg(a_hundred),steps=100-16      683µs ± 1%     686µs ± 4%    ~     (p=0.780 n=9+10)
RangeQuery/expr=avg(a_hundred),steps=1000-16    4.29ms ± 7%    4.34ms ± 2%    ~     (p=0.393 n=10+10)
```